+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.1.0 - 2019-03-31
+
+### Added
+
+- Support for .sass/.css loader in webpack
+- Support for production build of minified .css files
+
 ## 2.0.1 - 2019-03-31
 
 ### Added